move metric reporting function to crate::server
