+++ syncserver/src/server/mod.rs
@@ -1,7 +1,7 @@
 //! Main application server
 
 use std::{
-    fmt,
+    env, fmt,
     sync::{
         atomic::{AtomicU64, Ordering},
         Arc,
@@ -18,14 +18,14 @@ use actix_web::{
     middleware::errhandlers::ErrorHandlers,
     web, App, HttpRequest, HttpResponse, HttpServer,
 };
-use cadence::StatsdClient;
+use cadence::{Gauged, StatsdClient};
 use syncserver_common::InternalError;
-use syncserver_db_common::DbPool;
+use syncserver_db_common::{error::DbError, DbPool, GetPoolState, PoolState};
 use syncserver_settings::Settings;
 use syncstorage_settings::{Deadman, ServerLimits};
-use tokio::sync::RwLock;
+use tokio::{sync::RwLock, time};
 
-use crate::db::{pool_from_settings, spawn_pool_periodic_reporter};
+use crate::db::pool_from_settings;
 use crate::error::ApiError;
 use crate::server::metrics::Metrics;
 use crate::tokenserver;
@@ -275,7 +275,7 @@ impl Server {
                 blocking_threadpool.clone(),
             )?;
 
-            spawn_pool_periodic_reporter(
+            spawn_metric_periodic_reporter(
                 Duration::from_secs(10),
                 *state.metrics.clone(),
                 state.db_pool.clone(),
@@ -287,7 +287,7 @@ impl Server {
             None
         };
 
-        spawn_pool_periodic_reporter(
+        spawn_metric_periodic_reporter(
             Duration::from_secs(10),
             metrics.clone(),
             db_pool.clone(),
@@ -343,7 +343,7 @@ impl Server {
             blocking_threadpool.clone(),
         )?;
 
-        spawn_pool_periodic_reporter(
+        spawn_metric_periodic_reporter(
             Duration::from_secs(10),
             *tokenserver_state.metrics.clone(),
             tokenserver_state.db_pool.clone(),
@@ -366,7 +366,7 @@ impl Server {
     }
 }
 
-pub fn build_cors(settings: &Settings) -> Cors {
+fn build_cors(settings: &Settings) -> Cors {
     // Followed by the "official middleware" so they run first.
     // actix is getting increasingly tighter about CORS headers. Our server is
     // not a huge risk but does deliver XHR JSON content.
@@ -395,6 +395,55 @@ pub fn build_cors(settings: &Settings) -> Cors {
     }
 
     cors
+}
+
+/// Emit database pool and threadpool metrics periodically
+fn spawn_metric_periodic_reporter<T: GetPoolState + Send + 'static>(
+    interval: Duration,
+    metrics: StatsdClient,
+    pool: T,
+    blocking_threadpool: Arc<BlockingThreadpool>,
+) -> Result<(), DbError> {
+    let hostname = hostname::get()
+        .expect("Couldn't get hostname")
+        .into_string()
+        .expect("Couldn't get hostname");
+    let blocking_threadpool_size =
+        str::parse::<u64>(&env::var("ACTIX_THREADPOOL").unwrap()).unwrap();
+    tokio::spawn(async move {
+        loop {
+            let PoolState {
+                connections,
+                idle_connections,
+            } = pool.state();
+            metrics
+                .gauge_with_tags(
+                    "storage.pool.connections.active",
+                    (connections - idle_connections) as u64,
+                )
+                .with_tag("hostname", &hostname)
+                .send();
+            metrics
+                .gauge_with_tags("storage.pool.connections.idle", idle_connections as u64)
+                .with_tag("hostname", &hostname)
+                .send();
+
+            let active_threads = blocking_threadpool.active_threads();
+            let idle_threads = blocking_threadpool_size - active_threads;
+            metrics
+                .gauge_with_tags("blocking_threadpool.active", active_threads)
+                .with_tag("hostname", &hostname)
+                .send();
+            metrics
+                .gauge_with_tags("blocking_threadpool.idle", idle_threads)
+                .with_tag("hostname", &hostname)
+                .send();
+
+            time::delay_for(interval).await;
+        }
+    });
+
+    Ok(())
 }
 
 /// A threadpool on which callers can spawn non-CPU-bound tasks that block their thread (this is
@@ -432,7 +481,7 @@ impl BlockingThreadpool {
         result
     }
 
-    pub fn active_threads(&self) -> u64 {
+    fn active_threads(&self) -> u64 {
         self.spawned_tasks.load(Ordering::Relaxed)
     }
 }
